initial commit based on first prompt
I need a new personal portfolio website. I'm a software engineer with 19 years of experience. It needs to showcase the languages, frameworks, and platforms I've worked with, as well as examples of my work. There should also be a section for my select career accomplishments, career experience, and other side projects.